[*] Update to nom 5
